Catch errors when parsing JSON-LD
Sometimes there may be errors in the JSON provided on a web page. By
catching errors when parsing it, we can ignore any error that occurs.
Of course, if the error is in the recipe JSON-LD, we will not get the
recipe; but if the error occurs in different JSON-LD, we can still
extract the recipe.

Add a `org-chef-json-ld-debug' variable to surface errors when debugging.